% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_crosstabs_utils_helper.R
\name{gt_crosstab}
\alias{gt_crosstab}
\title{Creates a simple {gt} table for crosstabs.}
\usage{
gt_crosstab(data_, cols_)
}
\arguments{
\item{data_}{is the DRS dataset.}

\item{cols_}{is a character vector, length 2, of variable names, eg. Q1_0.}
}
\value{
object of class gt_tbl
}
\description{
Creates a simple {gt} table for crosstabs.
}
