% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graphs_fct.R
\name{mod_graph_prep}
\alias{mod_graph_prep}
\title{graph module plotting preparation function}
\usage{
mod_graph_prep(l)
}
\arguments{
\item{l}{list of tibbles. Each tibble contains a single question, but spans several rows due to each question. The tibble includes columns from 'shiny_parameters.xlsx'.}
}
\value{
list
}
\description{
graph module plotting preparation function
}
