% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycdrs-package.R
\docType{package}
\name{shinycdrs-package}
\alias{shinycdrs}
\alias{shinycdrs-package}
\title{shinycdrs: Shiny cdrs}
\description{
A shiny app to explore the California Delta Resident's Survey.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ktomari/shinycdrs}
  \item Report bugs at \url{https://github.com/ktomari/shinycdrs/issues}
}

}
\author{
\strong{Maintainer}: Kenji Tomari \email{ktomari@ucdavis.edu}

}
\keyword{internal}
